using long int

+++ ORBITER/src/lib/foundations/algebra_and_number_theory/algebra_and_number_theory.h
@@ -2193,8 +2193,8 @@ public:
 };
 
 
-void vector_space_unrank_point_callback(int *v, int rk, void *data);
-int vector_space_rank_point_callback(int *v, void *data);
+void vector_space_unrank_point_callback(int *v, long int rk, void *data);
+long int vector_space_rank_point_callback(int *v, void *data);
 
 }}
